stopwatch metrics for api/fe endpoints
