+++ requirements.txt
@@ -7,6 +7,7 @@ SQLAlchemy==0.9.1
 WTForms==2.0.1
 argparse==1.2.1
 backports.ssl-match-hostname==3.4.0.2
+dbx-stopwatch==1.0
 enum34==1.0.4
 expvar==0.0.2
 networkx==1.8.1
